renamed ConfirmationDialog -> PixelDialog
