+++ src/components/PixelDialog.vue
